Fix CI
The current setup fails because it can't find `brew`, even though Linux
homebrew is available on GitHub runners. It is however not on PATH by
default.

Instead of using homebrew on Linux, let's just use a nice action for
`shellcheck` and `go install` for `shfmt`. The latter does require some
fiddling with PATH, but at least this way you get the latest version.